requirements.txt
# Sabeel Homeo Clinic AI Chatbot - Dependencies

# FastAPI Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Configuration
python-dotenv>=1.0.0
pydantic-settings>=2.1.0

# Request parsing and async HTTP
python-multipart>=0.0.9
requests>=2.31.0
httpx>=0.27.0
aiohttp>=3.9.0

# WebSocket Support
websockets>=12.0

# AI & LLM
openai>=2.17.0

# Twilio (Voice & SMS)
twilio>=8.13.0

# Speech-to-Text
deepgram-sdk>=3.1.0

# Text-to-Speech
elevenlabs>=1.1.0

# Audio Processing
pydub>=0.25.1

# Supabase
supabase>=2.8.1
